requirements.txt
flask==1.0.2
numpy==1.16.2
pillow==5.4.0
shapely==1.7.0
pygdal==2.2.1.6
geopandas==0.6.2
scipy==1.4.1
rasterio==1.0.28
pandas==0.25.3
scikit-image==0.14.2
gunicorn
pyyaml
